Refactor number tests
